Update zenoss_gap_detection.py
Remodeled splitting scriptrole into scriptmode and scripttiming specifically to allow targeting back population to either the new ideas of 'late' and 'gap' with allowing 'both' (which actually showed a bug with point/pointrange where I wasn't doing the summary). Added some functionality for range, but didn't test it out, so can't be sure I did everything needed still haven't done anything for cleanup, but may have added some checks for it as a start, and error handling for range is extremely bad because it will fail when it tries to assign.

Overall it makes things much better by not mixing timing and what kind of data is being collected/posted/analyzed. And it finally settles on hopefully a better name for the old "present" and "historical" -- much better as "late" and "gap" with the only potential ambiguity of "late" is more significant when running as "realtime" and seems a little less clear if doing any of the non-realtime ones (including cleanup, which will run in real-time but will always do its work on something that isn't 'present time').