Update zenoss_gap_detection.py
To make it simpler for me to run I made the "relocation" of the creds.cfg file only kick in if "default" was the config set being used. Otherwise its going to look in the local folder. I probably could be more helpful to people, but this addresses my issue for now.