Update zenoss_gap_detection.py
Added a new option to detect that there's a huge gap and use the script start time (assuming it errored out enough to completely restart) to limit the back population of the gap. It shouldn't conflict with the other way of detecting that there's a lot of late items. If it works well "in the wild" I may get rid of the other large/numerous gap detection.

Also added some line feeds to shorten lines.