Update zenoss_gap_detection.py
Attempt to allow the specific situation of 'gap' running behind to be able to move forward only 300 seconds and continue on as if it was "point" until it needs to wait for the 5 minute interval again (e.g. it gets caught up). This may cause something to overlap...but that's not horrible either. Doesn't activate for iteration 0 (which shouldn't be delayed). Can't test this easily, both this and the last commit have to just be watched "in the wild".